install script config selection & fix semding empty file to groups

+++ scripts/install.sh
@@ -5,6 +5,7 @@ set -eu
 SYSTEMDDIR="/etc/systemd/system"
 MOONRAKER_BOT_ENV="${HOME}/moonraker-telegram-bot-env"
 MOONRAKER_BOT_DIR="${HOME}/moonraker-telegram-bot"
+MOONRAKER_BOT_LOG="${HOME}/klipper_logs/telegram.log"
 KLIPPER_CONF_DIR="${HOME}/klipper_config"
 CURRENT_USER=${USER}
 
@@ -24,6 +25,12 @@ init_config_path() {
     KLIPPER_CONF_DIR=${klipper_cfg_loc}
   fi
   report_status "Using configs from ${KLIPPER_CONF_DIR}"
+
+  report_status "Selecting log path"
+  echo -e "\n\n\n"
+  read -p "Enter your bot log file: " -e -i "${MOONRAKER_BOT_LOG}" bot_log_path
+  MOONRAKER_BOT_LOG=${bot_log_path}
+  report_status "Writing bot logs to ${MOONRAKER_BOT_LOG}"
 }
 
 stop_sevice() {
@@ -68,9 +75,13 @@ create_service() {
     mv "${KLIPPER_CONF_DIR}"/application.conf "${KLIPPER_CONF_DIR}"/telegram.conf
   fi
 
-  report_status "Creating base config file"
-  cp -n "${MOONRAKER_BOT_DIR}"/scripts/base_install_template "${KLIPPER_CONF_DIR}"/telegram.conf
-  cp -rn "${MOONRAKER_BOT_DIR}"/imgs "${KLIPPER_CONF_DIR}"/
+  if [[ ! -f "${KLIPPER_CONF_DIR}"/telegram.conf ]]; then
+    report_status "Creating base config file"
+    cp -n "${MOONRAKER_BOT_DIR}"/scripts/base_install_template "${KLIPPER_CONF_DIR}"/telegram.conf
+    cp -rn "${MOONRAKER_BOT_DIR}"/imgs "${KLIPPER_CONF_DIR}"/
+
+    sed -i "s+some_log_path+${MOONRAKER_BOT_LOG}+g" "${KLIPPER_CONF_DIR}"/telegram.conf
+  fi
 
   ### create systemd service file
   sudo /bin/sh -c "cat > ${SYSTEMDDIR}/moonraker-telegram-bot.service" <<EOF
